Correction de bug mineur

+++ dessin_de_courbes/dessin_de_courbes/drawMargin.c
@@ -19,14 +19,14 @@ height:	Hauteur totale de l'image
 
 #include "typedef.h"
 
-void drawMargin (unsigned int margin, sRGB ** data, unsigned int width, unsigned int height)
+void drawMargin(unsigned int margin, sRGB ** data, unsigned long width, unsigned long height)
 {
 	sRGB grey = { 0xCD, 0xCD, 0xCD };
 
 	//	Y coordinate manipulation. Bitmap is read from last to first line
-	for (unsigned y = 0; y < height; y++)
+	for (unsigned long y = 0; y < height; y++)
 	{
-		for (unsigned x = 0; x < width; x++)
+		for (unsigned long x = 0; x < width; x++)
 		{
 			if (((x == margin || x == width - margin) && y>margin && y < height - margin) || ((y == margin || y == height - margin) && x > margin && x < width - margin))
 			{
